pyproject.toml
[tool.poetry]
name = "software-inventory-collector"
version = "0.1.0"
description = "Client collecting data exported by https://github.com/dparv/inventory-exporter"
authors = ["Diko Parvanov <diko.parvanov@canonical.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
packages = [{include = "software_inventory_collector"}]

[tool.poetry.dependencies]
python = "^3.8.1"
requests = "^2.28.2"
PyYAML = "^6.0"
juju = "2.9.38.1"

[tool.poetry.scripts]
software-inventory-collector = "software_inventory_collector.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.coverage.run]
source = ["software_inventory_collector"]

[tool.flake8]
ignore = []
exclude = ['.eggs', '.git', '.tox', '.venv', '.build', 'report']
max-line-length = 90
max-complexity = 10
per-file-ignores = "__init__.py:F401"

[tool.black]
line-length = 90
exclude = '''
/(
    | .eggs
    | .git
    | .tox
    | .venv
    | .build
    | report
)/
'''

[tool.isort]
profile = "black"
skip_glob = '''
/(
    | .eggs
    | .git
    | .tox
    | .venv
    | .build
    | report
)/
'''

[tool.mypy]
warn_unused_ignores = true
warn_unused_configs = true
warn_unreachable = true
disallow_untyped_defs = true
exclude = ['.eggs', '.git', '.tox', '.venv', '.build', 'report', 'tests']

## Ignore unsupported imports
[[tool.mypy.overrides]]
module = [
  "juju.*",
]
ignore_missing_imports = true
